dlp_worker/requirements.txt
aiohttp==3.9.5
aiobotocore==2.13.0
python-dotenv==1.0.1
django==4.2